requirements.txt
Django>=3.2,<4.0
requests
djangorestframework
python-dotenv
psycopg2-binary>=2.8
